### 2019-06-24 #1522 @whedon
[PRE REVIEW]: Singularity Compose: Orchestration for Singularity Instances
**Submitting author:** @vsoch (<a href="http://orcid.org/0000-0002-4387-3819">Vanessa Sochat</a>)
**Repository:** <a href="https://github.com/singularityhub/singularity-compose" target ="_blank">https://github.com/singularityhub/singularity-compose</a>
**Version:** 0.0.15
