### 2019-06-24 #1522 @kyleniemeyer
Hi @vsoch, thanks for your submission. At first glance, while this looks like valuable software, I don't see anything that indicates that it is *research* software—nothing in the README or paper seems to connect it to research. Could you clarify how it has an obvious research application?
