### 2019-06-24 #1522 @vsoch
Think of any kind of research application that involves a database (mongodb, postgres, etc.) along with an application to interact with it. For example, Singularity Compose is going to fit in nicely with being used on our SLURM cluster to start a database on a node for a researcher's (containerized) scripts to interact with. A second example is with respect to interactive applications. I'm not a huge fan, but jupyter notebooks (and similar) and becoming more popular, and tools like OnDemand (Ohio Supercomputing Center) now allow for running rstudio/jupyter via a Singularity container.  An integration with Singularity compose would allow for a user to do similar, but with the interactive applications and services that are relevant for their work.

The use cases are strong with this one, and there is no tool that can provide what I outlined above. It is niche as it's for Singularity containers, however they are the unit of reproducible operability on HPC. You are correct that it's a general purpose tool, and not something like a package for a specific scientific domain. But the reasons that I outlined above, and what can be developed for research with such a tool, are exactly what makes it so exciting. I've been waiting years (since 2016) to develop this, and the base unit (Singularity) is finally ready to support it.
