### 2019-06-24 #1522 @vsoch
And of course reproducibility of all of the above - given that I run my specific database service + application, by placing the singularity-compose.yml alongside the version controlled code and container recipes, this ensures not just reproducibility of the container build, but also deployment of the services to do some analysis. I believe I stated this in the paper. :)
