### 2019-06-24 #1522 @vsoch
@kyleniemeyer please let me know if you have other questions, or if you'd like me to suggest reviewers (those with experience with reproducibility and/or container technologies would be well suited!)
