### 2019-06-24 #1522 @vsoch
oh that's so cool! @whedon extracts labels now! You're the best robot, @whedon  <3
