### 2019-06-24 #1522 @kyleniemeyer
@vsoch great, thank you for the information. I do think that both the README and JOSS paper need to emphasize the research application a bit more, so it's clear, but that can happen through the review process.
