### 2019-06-27 #1522 @gkthiruvathukal
@vsoch Can you help suggest a a few reviewers from [this list of people](https://bit.ly/joss-reviewers)? We now require 2 reviewers for all JOSS submissions.
