### 2019-06-27 #1522 @vsoch
Sure thing! It looks like @nuest is the only other one that lists containers, so he might be familiar with Docker Compose (Singularity Compose's counterpart). The next tag that might imply container experience would be reproducibility, and from that I can suggest @cole-brokamp and @daissi.
