### 2019-06-27 #1522 @cole-brokamp
I am out traveling for the next three weeks, but would be glad to take a look when I get back.

> On Jun 27, 2019, at 3:18 PM, George K. Thiruvathukal <notifications@github.com> wrote:
>
> @nuest, @cole-brokamp, @daissi: Would you be willing to contribute a review for this JOSS submission? I need two reviewers to begin review.
>
