### 2019-06-27 #1522 @cole-brokamp
Sure, sounds good!

> On Jun 27, 2019, at 3:32 PM, Vanessasaurus <notifications@github.com> wrote:
>
> Works for me too! Safe travels @cole-brokamp ✈️
>
