### 2019-07-18 #1578 @vsoch
Thanks whedon! The content of the paper looks good, but there are "nd-(a-c)" that likely I can fix by adding a date.
