### 2019-07-18 004928e
try adding date to software references
Signed-off-by: Vanessa Sochat <vsochat@stanford.edu>
