+++ paper/paper.bib
@@ -33,7 +33,8 @@
   author       = "Sochat, Vanessa",
   abstract     = "orchestration tool for Singularity container instances",
   howpublished = "\url{https://singularityhub.github.io/singularity-compose}",
-  note         = "Accessed: 2019-6-24"
+  note         = "Accessed: 2019-6-24",
+  date         = "2019-06-24"
 }
 
 
@@ -43,7 +44,8 @@
   author       = "Sochat, Vanessa",
   abstract     = "open source code for singularity-compose",
   howpublished = "\url{https://github.com/singularityhub/singularity-compose}",
-  note         = "Accessed: 2019-6-24"
+  note         = "Accessed: 2019-6-24",
+  date         = "2019-06-24"
 }
 
 
@@ -65,7 +67,8 @@
   author       = "Sochat, Vanessa",
   abstract     = "open source examples for singularity-compose",
   howpublished = "\url{https://github.com/singularityhub/singularity-compose-examples}",
-  note         = "Accessed: 2019-6-24"
+  note         = "Accessed: 2019-6-24",
+  date         = "2019-06-24"
 }
 
 @MISC{Merkel2014-da,
