+++ paper/paper.bib
@@ -110,7 +110,8 @@
 @MISC{SingularityInstances,
   title        = "Running Services --- Singularity container 3.2 documentation",
   howpublished = "\url{https://sylabs.io/guides/3.2/user-guide/running_services.html?highlight=instances}",
-  note         = "Accessed: 2019-6-24"
+  note         = "Accessed: 2019-6-24",
+  year         = 2019
 }
 
 
