### 2019-07-23 #1578 @gkthiruvathukal
@cole-brokamp Hmm, I thought I assigned you as a reviewer but you are not showing up on the list of reviewers (and therefore don't have a checklist). I am not sure why you're not showing up.  I am going to see if someone from @openjournals/joss-eics can add you.

In the meantime, you can go ahead and list any issues here as I definitely want you involved in this review.
