### 2019-08-01 singularityhub/singularity-compose#22 @cole-brokamp
[JOSS review] error when running singularity-compose-simple example
Hello, I am using a docker container (ubuntu:18.04) to test the `singularity-compose-simple` example as in the Quick Start documentation.  I've installed singularity from source (version 2.5.2) and successfully installed `singularity-compose` with `pip`, but get the following error whenever I try to run any command, for example `singularity-compose build` within the `singularity-compose-simple` folder from the example repository:

```sh
> singularity-compose build
Traceback (most recent call last):
  File "/usr/local/bin/singularity-compose", line 7, in <module>
    from scompose.client import start
  File "/usr/local/lib/python2.7/dist-packages/scompose/client/__init__.py", line 198
    from .exec import main
             ^
SyntaxError: invalid syntax
```

Any ideas on where I could have gone wrong?
