### 2019-08-01 singularityhub/singularity-compose#22 @vsoch
There are two quick fixes here. The first is that singularity-compose won't work with python 2.x, I just can't provide support for it given end of life.

The second issue (which you haven't hit yet) is that Singularity 2.x definitely won't support instances like we need. The reason we needed to wait for the 3.x family to create this, period, is because it had very little support for networking before that.

So unfortunately to test you need both of those things updated.
