### 2019-08-01 singularityhub/singularity-compose#22 @vsoch
You also likely need to test this on a host - I haven't tried in a docker container, but given the requirements for networking, I suspect it won't work.
