### 2019-08-01 bdcff37
adding dependencies section
Signed-off-by: Vanessa Sochat <vsochat@stanford.edu>
