### 2019-08-01 singularityhub/singularity-compose#22 @vsoch
I realize this isn't in the docs clearly stated! Just added here https://singularityhub.github.io/singularity-compose/#/?id=dependencies (open devtools to clear cache if you don't see it yet).
