### 2019-08-01 singularityhub/singularity-compose#22 @cole-brokamp
Perfect, you beat me to my own suggestion! This is super helpful for a beginner with this type of setup for me.  (I usually rely on our HPC version of Singularity, which they haven't yet upgraded to 3.0).
