### 2019-08-01 singularityhub/singularity-compose#22 @cole-brokamp
I think you are right about not working in a docker container, I'm getting an `ERROR: Failed to unshare root file system: Operation not permitted`, which I'm pretty sure would have to do with how Docker manages that.  I will try on a host this evening. Thanks for the help!
