### 2019-08-01 singularityhub/singularity-compose#22 @vsoch
I haven't tried it, but you would need to set privileged to true, and then bind the docker daemon, as is done with docker2singularity - [see the docker run example here](https://github.com/singularityware/docker2singularity#build-a-squashfs-image). I highly *don't* recommend this - it's masterfully messed up my computer when I've tried in random cases.
