### 2019-08-01 singularityhub/singularity-compose#22 @cole-brokamp
Going to stay away from that, then!  I found out that our HPC recently had singularity 3.0 installed.  So I'm giving it a go from there right now.  Do I need to be a sudoer to run `singularity-compose build`?  Its complaining to me that there is an issue building the container.
