### 2019-08-01 singularityhub/singularity-compose#22 @vsoch
okay - but if it's 3.0 exactly that's too early for a fully working example - likely a single container example will work, but not if they need to talk to one another. The same rules are true for Singularity - you can't build without sudo on a shared resource. You can try adding the fakeroot option to the file, but that's only available with 3.3.0 and later. So - I think you will need to test this on your local machine, because the cluster version is still too early :/
