### 2019-08-01 singularityhub/singularity-compose#22 @cole-brokamp
OK, thanks.  They are on 3.1. I'll try at home this evening.
