### 2019-08-01 #1578 @cole-brokamp
FYI... @vsoch is currently helping me through some of my own issues with running the example (totally my own fault!) over here: https://github.com/singularityhub/singularity-compose/issues/22

...I'll report back soon
