### 2019-08-01 #1578 @vsoch
Totally not his fault! Singularity changes *so* fast that nobody can really keep up.
