### 2019-08-01 #1578 @gkthiruvathukal
Thanks, @cole-brokamp and all. Just wanted to get an idea of where we stand.
