### 2019-08-02 singularityhub/singularity-compose#22 @cole-brokamp
all good with Singularity 3.3.0
