### 2019-08-25 #1578 @gkthiruvathukal
@cole-brokamp and @daissi, Thanks for the help in reviewing this submission.

@vsoch, I am ready to move toward acceptance. Please do the following:
