### 2019-08-25 #1578 @vsoch
- [x] The tagged release is for version 0.0.16, available [here](https://github.com/singularityhub/singularity-compose/releases/tag/0.0.16)
 - [x] The archive on Zenodo is [here](https://zenodo.org/record/3376793#.XWLGkHFKhgw)
 - [x] For metadata, I made sure my full name was used, affiliation, orcid id, the title corresponds with the name of the paper, and the release is 0.0.16
 - [x] The zenodo DOI (along with other metadata) is provided below!

```
Vanessa Sochat. (2019, August 25). Singularity Compose: Orchestration for Singularity Instances (Version 0.0.16). Zenodo. http://doi.org/10.5281/zenodo.3376793
```
I updated the version in the first comment here too.
