### 2019-08-25 #1578 @gkthiruvathukal
Hi @vsoch Can you take a look at the invalid DOIs in the references? I think the problem is that you don't have the full DOI: 10.1371/journal.pone.0177459
