+++ paper/paper.bib
@@ -17,7 +17,7 @@
                previously unmet need, makes Singularity a game changing
                development for computational science.",
   journal   = "PLoS One",
-  doi       = "journal.pone.0177459",
+  doi       = "https://doi.org/10.1371/journal.pone.0177459",
   publisher = "Public Library of Science",
   volume    =  12,
   number    =  5,
@@ -50,14 +50,14 @@
 
 
 
-@MISC{DockerCompose,
+@Software{DockerCompose,
   title        = "Docker Compose",
   booktitle    = "Docker Documentation",
+  author       = "Docker, Inc.",
   abstract     = "Introduction and Overview of Compose",
-  month        =  jun,
-  year         =  2019,
   howpublished = "\url{https://docs.docker.com/compose/}",
   note         = "Accessed: 2019-6-24"
+  date         = "2019-06-24"
 }
 
 
