+++ paper/paper.bib
@@ -56,7 +56,7 @@
   author       = "Docker, Inc.",
   abstract     = "Introduction and Overview of Compose",
   howpublished = "\url{https://docs.docker.com/compose/}",
-  note         = "Accessed: 2019-6-24"
+  note         = "Accessed: 2019-6-24",
   date         = "2019-06-24"
 }
 
