### 2019-08-25 #1578 @vsoch
The Docker Compose entry shouldn't have a DOI, it's for a software page (so I think safe to ignore).
