### 2019-08-25 #1578 @gkthiruvathukal
@vsoch Thanks for your follow up. I'm ready to move forward with accepting your submission.

@openjournals/joss-editors I recommend acceptance, pending any additional input you may have.
