### 2019-08-25 #1578 @kthyng
@gkthiruvathukal You want to use @openjournals/joss-eics here!
