### 2019-08-26 #1578 @kyleniemeyer
Hi @vsoch, just a few edits needed on the paper:
- Could you add more-complete affiliation info?
- For the "Running services — singularity container 3.2 documentation." reference, I think you can add "Sylabs Inc" as the author
- For the Kurtzer reference, could you strip the "https://doi.org/" from the DOI in the bib file? Our template only expects the DOI itself
