### 2019-08-26 #1578 @vsoch
Could you clarify where you would like the affiliation info? I have it in the paper and zenodo record.

For the second point, I'm happy to do this, but for historical correctness I need to point out that the majority of content for that page was copy pasted verbatim from https://github.com/singularityware/singularityware.github.io/blob/master/pages/docs/user-docs/docs-instances.md, which was written well before the existence of Sylabs, by community members too.
