### 2019-08-26 9f97f24
adding author to instances docs, and stripping https from doi
Signed-off-by: Vanessa Sochat <vsochat@stanford.edu>
