+++ paper/paper.bib
@@ -17,7 +17,7 @@
                previously unmet need, makes Singularity a game changing
                development for computational science.",
   journal   = "PLoS One",
-  doi       = "https://doi.org/10.1371/journal.pone.0177459",
+  doi       = "10.1371/journal.pone.0177459",
   publisher = "Public Library of Science",
   volume    =  12,
   number    =  5,
@@ -109,6 +109,7 @@
 
 @MISC{SingularityInstances,
   title        = "Running Services --- Singularity container 3.2 documentation",
+  author       = "Singularity contributors",
   howpublished = "\url{https://sylabs.io/guides/3.2/user-guide/running_services.html?highlight=instances}",
   note         = "Accessed: 2019-6-24",
   year         = 2019
