### 2019-08-26 #1578 @vsoch
Changes made - I chose "Singularity contributors" to the authors to be more correct, and stripped the DOI. I will wait for feedback about the affiliation.
