### 2019-08-26 #1578 @vsoch
I changed "Stanford University" to "Stanford University Research Computing Center" in case that is what you are wanting.
