### 2019-08-26 #1578 @kyleniemeyer
@vsoch For the affiliation, I just want more-complete info, like what you'd see on a typical article: department, institution, city, state, country (or similar).

For that reference, the choice of author is fine, I just wanted someone listed in the author field.
