### 2019-08-26 #1578 @kyleniemeyer
@vsoch (for the affiliation field in the paper)
