### 2019-08-26 09c9391
adding a bunch of stanfords to the affiliation
Signed-off-by: Vanessa Sochat <vsochat@stanford.edu>
