### 2019-08-26 #1578 @vsoch
okay just added a bunch of Stanfords to it :)
