### 2019-08-26 #1578 @vsoch
Just a comment - the example paper only has the university title for the affiliation, https://joss.readthedocs.io/en/latest/submitting.html?highlight=affiliation#example-paper-and-bibliography which is what I've done for the last (maybe 6?) papers. If the requirement is to have these extraneous fields, it should be clearly stated... somewhere, and then the criteria checked consistent across reviews.
