### 2019-08-26 ae23ee6
adding joss badge
Signed-off-by: Vanessa Sochat <vsochat@stanford.edu>
