+++ README.md
@@ -1,5 +1,6 @@
 # Singularity Compose
 
+[![status](http://joss.theoj.org/papers/1fc2593b11b5e18df12efb59ed8757a0/status.svg)](http://joss.theoj.org/papers/1fc2593b11b5e18df12efb59ed8757a0)
 [![DOI](https://zenodo.org/badge/188852712.svg)](https://zenodo.org/badge/latestdoi/188852712)
 
 This is a simple orchestration library for Singularity containers, akin to
